added user page functionality + styling added authentication guard added media query for header --> todo: add hamburger menu or similar

+++ BoulderSite/src/app/Components/user/user.component.ts
@@ -1,5 +1,8 @@
-import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
+import { Component, OnDestroy, OnInit } from '@angular/core';
+import { FormBuilder, Validators } from '@angular/forms';
+import { ActivatedRoute, Router } from '@angular/router';
+import { Subscription } from 'rxjs';
+import { Climb } from 'src/app/shared/models/climb.model';
 import { User } from 'src/app/shared/models/user.model';
 import { ClimbService } from 'src/app/shared/services/climb.service';
 import { CommentService } from 'src/app/shared/services/comment.service';
@@ -13,9 +16,18 @@ import { UserService } from 'src/app/shared/services/user.service';
   styleUrls: ['./user.component.sass']
 })
 
-export class UserComponent {
+export class UserComponent implements OnInit, OnDestroy {
+  id: number = 0;
   imageSrc: string = '';
   user: User = new User();
+  climbs: Climb[] = new Array();
+  flashPercent: number = 0;
+  selectedAction: number = 0;
+  showPassword: boolean = false;
+  updatePwForm = this.fb.group({
+    newPassword: ['', [Validators.required, Validators.minLength(8), Validators.pattern('[a-zA-Z!-@]*')]]
+  });
+  routerSubscription?: Subscription;
 
   constructor(
     private commentService: CommentService,
@@ -23,18 +35,46 @@ export class UserComponent {
     private ratingService: RatingService,
     public userService: UserService,
     public imageService: ImageService,
-    private router: Router) {
-    const id = this.userService.getUserId();
-    this.userService.getById(id).subscribe({
-      next: (data: User) => {
+    private activatedRoute: ActivatedRoute,
+    private router: Router,
+    private fb: FormBuilder) { }
+  ngOnInit(): void {
+    this.routerSubscription = this.activatedRoute.params.subscribe(p => {
+      this.id = Number.parseInt(p['id']);
+      // get user
+      this.userService.getById(this.id).subscribe((data: User) => {
         this.user = data;
         if (this.user.picture == null) {
-          this.imageSrc = 'assets/img/bob.jpg'
+          this.imageSrc = 'assets/img/defaultUser.jpg'
         } else {
           this.imageSrc = 'http://localhost:9000/storage/img/users/' + this.user.picture;
         }
-      }
-    })
+      });
+
+      // get climbs
+      this.climbService.getByUser(this.id).subscribe(data => this.climbs = data);
+      // get flashes
+      const flashes = this.climbs.filter(climb => {
+        return climb.isFlash == true;
+      });
+      this.flashPercent = Math.floor((flashes.length / this.climbs.length) * 100);
+    });
+  }
+
+  ngOnDestroy(): void {
+    this.routerSubscription?.unsubscribe();
+  }
+
+  setSelectedAction(a: number) {
+    if (this.selectedAction === a) {
+      this.selectedAction = 0;
+    } else {
+      this.selectedAction = a;
+    }
+  }
+
+  toggleShowPassword() {
+    this.showPassword = !this.showPassword;
   }
 
   updateUserInfo() {
@@ -66,9 +106,11 @@ export class UserComponent {
   }
 
   deleteImage() {
-    this.imageService.delete('users', this.user.userId, this.user.picture);
-    this.user.picture = '';
-    this.imageSrc = 'assets/img/bob.jpg'
+    if (this.user.picture != null) {
+      this.imageService.delete('users', this.user.userId, this.user.picture);
+      this.user.picture = null;
+      this.imageSrc = 'assets/img/defaultUser.jpg'
+    }
   }
 
   deleteUser() {
